+++ src/main/java/cart/product/web/dto/ProductAdminViewResponse.java
@@ -9,9 +9,9 @@ public class ProductAdminViewResponse {
     private final Long id;
     private final String name;
     private final String image;
-    private final Integer price;
+    private final Long price;
 
-    public ProductAdminViewResponse(Long id, String name, String image, Integer price) {
+    public ProductAdminViewResponse(Long id, String name, String image, Long price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -44,7 +44,7 @@ public class ProductAdminViewResponse {
         return image;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
 }
