+++ src/main/java/cart/product/web/dto/ProductHomeViewResponse.java
@@ -9,9 +9,9 @@ public class ProductHomeViewResponse {
     private final Long id;
     private final String name;
     private final String image;
-    private final Integer price;
+    private final Long price;
 
-    public ProductHomeViewResponse(Long id, String name, String image, Integer price) {
+    public ProductHomeViewResponse(Long id, String name, String image, Long price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -44,7 +44,7 @@ public class ProductHomeViewResponse {
         return image;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
 }
