+++ src/main/java/cart/product/web/dto/ProductResponse.java
@@ -9,9 +9,9 @@ public class ProductResponse {
     private final Long id;
     private final String name;
     private final String image;
-    private final Integer price;
+    private final Long price;
 
-    public ProductResponse(Long id, String name, String image, Integer price) {
+    public ProductResponse(Long id, String name, String image, Long price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -44,7 +44,7 @@ public class ProductResponse {
         return image;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
